requirements.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coverage==7.6.8
decorator==5.1.1
Flask==3.1.0
Flask-Cors==5.0.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
moviepy==2.1.1
numpy==2.1.3
packaging==24.2
pillow==10.4.0
pluggy==1.5.0
proglog==0.1.10
pytest==8.3.3
pytest-cov==6.0.0
python-dotenv==1.0.1
pytube==15.0.0
requests==2.32.3
setuptools==75.6.0
tqdm==4.67.1
urllib3==2.2.3
Werkzeug==3.1.3
yt-dlp==2024.11.18